Using Result pattern (#33)
* removing users.common

* adding mass transit package

* removing contracts

* domain events and mass transit initial setup

* using result pattern, primary constructor on user repository

* oopsies

* removed unused package

* removing unused packages

* return bad request on update user endpoint

+++ src/Users.Application/Users/Commands/DeleteUser/DeleteUserCommandHandler.cs
@@ -1,18 +1,25 @@
+using MassTransit;
 using SharedKernal.Messaging;
+using SharedKernal.Results;
 using Users.Application.Abstractions;
 
 namespace Users.Application.Users.Commands.DeleteUser;
 
 internal sealed class DeleteUserCommandHandler(IUserRepository userRepository)
-    : ICommandHandler<DeleteUserCommand>
+    : ICommandHandler<DeleteUserCommand, Result<bool>>
 {
-    public async Task Handle(DeleteUserCommand request, CancellationToken cancellationToken)
+    public async Task<Result<bool>> Handle(DeleteUserCommand request, CancellationToken cancellationToken)
     {
-        var user = await userRepository.GetByIdAsync(request.Id, cancellationToken)
-            ?? throw new KeyNotFoundException($"User '{request.Id}' was not found.");
+        var user = await userRepository.GetByIdAsync(request.Id, cancellationToken);
+
+        if (user is null)
+        {
+            return Error.NotFound(ErrorCodes.NotFound, $"user with {request.Id} not found");
+        }
 
         user.Delete();
         userRepository.Remove(user);
         await userRepository.SaveChangesAsync(cancellationToken);
+        return true;
     }
 }
